teacher/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
httpx==0.27.2
PyYAML==6.0.2
paho-mqtt==2.1.0
